requirements.txt
# ============================================================
# UV Skincare Advisor — Python Dependencies
# Class 12 Capstone Project
# ============================================================
# Install all at once:
#   pip install -r requirements.txt
# ============================================================

# --- Core Web Framework ---
streamlit==1.35.0     # Builds the interactive web app

# --- API & Networking ---
requests==2.31.0      # HTTP calls to Open-Meteo + Gemini APIs

# --- Data Handling ---
pandas==2.2.2         # Data tables

# --- Visualization ---
plotly==5.22.0        # Interactive charts
altair==5.3.0         # Declarative charts

# --- Image Handling ---
Pillow==10.3.0        # Skin tone visual reference images

# --- AI Ingredient Scanner ---
python-dotenv==1.0.1  # Loads GEMINI_API_KEY from .env file safely